gabino/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
ez-setup==0.9
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2-binary==2.9.3
pytz==2022.2.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
zipp==3.8.1
